+++ CHANGELOG.md
@@ -40,6 +40,7 @@ The **parallel** update.
 * Adding `xan freq -p/--parallel & -t/--threads`.
 * Adding `xan stats -p/--parallel & -t/--threads`.
 * Adding `xan search -p/--parallel & -t/--threads`.
+* Adding `xan agg -p/--parallel & -t/--threads`.
 * Adding piped column access to expression given to `xan flatmap -r`.
 * Adding `xan rename -R/--replace` & `xan rename -x/--suffix`.
 * Adding `xan parallel freq -l/--limit, -A/--all, -a/--approx & -N/--no-extra`.